+++ qu3d/cosmology_3d.hpp
@@ -150,7 +150,7 @@ namespace fidcosmo {
             int N = lnk.size();
             double dlnk = lnk[1] - lnk[0];
 
-            for (size_t i = 1; i < N - 1; ++i)
+            for (int i = 1; i < N - 1; ++i)
                 if (fabs(lnk[i + 1] - lnk[i] - dlnk) > 1e-8)
                     throw std::runtime_error(
                         "Input PlinearFilename does not have equal ln k spacing.");
